add QUIT command

+++ srcs/Server/Server.cpp
@@ -40,7 +40,7 @@ void Server::start()
 			_context->clearEmptyData();
 		} catch (std::bad_alloc& e) {
 			cout << "Allocation failed " << e.what() << endl;
-		} catch (std::exception &e) {
+		} catch (std::exception& e) {
 			cout << "Unexpected error " << e.what() << endl;
 		}
 	}
@@ -48,8 +48,10 @@ void Server::start()
 
 void Server::polling()
 {
-	if (poll(&(_pollfds.front()), _pollfds.size(), POLL_TIMEOUT) < 0)
-		cerr << "  poll() failed: " << strerror(errno) << endl;
+	if (poll(&(_pollfds.front()), _pollfds.size(), POLL_TIMEOUT) < 0) {
+		if (errno != EWOULDBLOCK)
+			cerr << "  poll() failed: " << strerror(errno) << endl;
+	}
 }
 
 void Server::_handlePolls()
@@ -91,7 +93,7 @@ bool Server::receiveMessage(User* user)
 	cout << "User listens:" << endl;
 	bytesRead = recv(user->getFD(), buffer, sizeof(buffer), 0);
 
-	if (bytesRead == 0) {
+	if (bytesRead <= 0 && errno != EWOULDBLOCK) {
 		cout << "Client ended the _userfd!" << user->getFD() << endl;
 		close(user->getFD());
 		_context->deleteUser(user);
